requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
firebase-admin==6.2.0
google-generativeai==0.3.2
meilisearch==0.28.0
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2 